requirements.txt
alembic==1.8.0
Babel==2.10.3
blinker==1.4
click==8.1.3
config==0.5.1
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-BabelEx==0.9.4
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
speaklater==1.3
SQLAlchemy==1.4.39
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
